+++ Changelog.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning].
 
 
 
+## [0.0.2] - 2024-07-14
+
+### Added
+
+- Chatlog history is now loadable!
+If there is atleast a single log before the standard 10 last logs there will be a message at the top signaling there are more logs to be loaded on demand.
+
+- A tooltip for each player has been added, includes right now:
+    - Class
+    - Race
+
+    If RaiderIO is installed and there is data for this player:
+    
+    - M+ Rating
+    - Raid progress
+
+### Fixed
+
+- Chat line height of status or blank lines should be consistent now.
+
+
 ## [0.0.1] - 2024-07-13
 
 ### Added